#34 Feat: 게시글 작성 시 뮤지컬 용어 함께 반환

+++ src/main/java/encore/server/domain/comment/repository/CommentRepository.java
@@ -15,7 +15,7 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
     Optional<Comment> findByIdAndPostAndDeletedAtIsNull(Long id, Post post);
 
     List<Comment> findAllByPostAndDeletedAtIsNull(Post post);
-    Integer countByPostAndDeletedAtIsNull(Post post);
+    Long countByPostAndDeletedAtIsNull(Post post);
 
     @Modifying(clearAutomatically = true)
     @Query("UPDATE Comment c SET c.deletedAt = CURRENT_TIMESTAMP WHERE c.post.id = :postId")
